requirements.txt
connexion==2.14.1
Flask==2.2.2
flask_marshmallow==0.14.0
flask_sqlalchemy==3.0.2
mysql_connector_repackaged==0.3.1
python-dotenv==0.21.0
